OSX 10.5 API fixups for mingw
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@28092 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSFileManager.m
@@ -3086,7 +3086,7 @@ static NSSet	*fileKeys = nil;
   return [NSDate dateWithTimeIntervalSince1970: statbuf.st_mtime];
 }
 
-- (unsigned long) filePosixPermissions
+- (NSUInteger) filePosixPermissions
 {
   return (statbuf.st_mode & ~S_IFMT);
 }
@@ -3223,12 +3223,12 @@ static NSSet	*fileKeys = nil;
   return statbuf.st_size;
 }
 
-- (unsigned long) fileSystemFileNumber
+- (NSUInteger) fileSystemFileNumber
 {
   return statbuf.st_ino;
 }
 
-- (unsigned long) fileSystemNumber
+- (NSUInteger) fileSystemNumber
 {
   return statbuf.st_dev;
 }
